pyproject.toml
[project]
name = "shortchain"
version = "0.1.0"
description = "A lightweight agent framework inspired by LangChain"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "openai>=1.0",
    "python-dotenv>=1.0",
    "pydantic>=2.0",
    "mcp>=1.0",
]
